Arrange files more neatly
